+++ spec/factories/topics.rb
@@ -1,6 +1,8 @@
 FactoryBot.define do
   factory :topic do
-    name { "MyString" }
-    description { "MyText" }
+    name { Faker::Educator.course_name }
+    description { Faker::Religion::Bible.quote }
+    association :subject
+    image_url { Faker::Internet.url }
   end
 end